-The downloadFileAsync is a lie, we actually want DownloadFileTaskAsync -Moved services to the project they are consumed in
