+++ Launcher/Launcher.csproj
@@ -32,11 +32,9 @@
 		<PackageReference Include="Octokit" Version="0.50.0" />
 		<PackageReference Include="Microsoft.DotNet.UpgradeAssistant.Extensions.Default.Analyzers" Version="0.2.212405">
 			<PrivateAssets>all</PrivateAssets>
-		</PackageReference>
+		</PackageReference>	
 		<!-- Only needed for net core 3.1+ -->
 		<!-- <PackageReference Include="Microsoft.Windows.Compatibility" Version="5.0.2" /> -->
 	</ItemGroup>
-	<ItemGroup>
-	  <ProjectReference Include="..\Syn3Updater\Syn3Updater.csproj" />
-	</ItemGroup>
+	<Import Project="..\SharedCode\SharedCode.projitems" Label="Shared" />
 </Project>
