+++ SharedCode/SharedCode.projitems
@@ -9,8 +9,6 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Classes\LauncherPrefs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Methods\JsonHelpers.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Services\DownloadViewModelService.cs" />
-    <Compile Include="..\SharedCode\Services\HomeViewModelService.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="$(MSBuildThisFileDirectory)Classes\" />
