+++ Syn3Updater/UI/Tabs/DownloadViewModel.cs
@@ -10,7 +10,7 @@ using System.Threading.Tasks;
 using System.Windows;
 using Cyanlabs.Syn3Updater.Helper;
 using Cyanlabs.Syn3Updater.Model;
-using Cyanlabs.Updater.Common;
+using Cyanlabs.Updater.Services;
 
 namespace Cyanlabs.Syn3Updater.UI.Tabs
 {
